refactor to address issue #93
