+++ multiSero/array_analyzer/extract/constants.py
