+++ multiSero/array_analyzer/utils/io_utils.py
