+++ multiSero/array_analyzer/workflows/interpolation_wf.py
@@ -4,16 +4,16 @@ import numpy as np
 import pandas as pd
 import skimage.io as io
 
-import array_analyzer.extract.image_parser as image_parser
-import array_analyzer.extract.txt_parser as txt_parser
-import array_analyzer.extract.img_processing as img_processing
-import array_analyzer.load.debug_plots as debug_plots
-import array_analyzer.load.report as report
-import array_analyzer.extract.constants as constants
-import array_analyzer.transform.array_generation as array_gen
-import array_analyzer.extract.background_estimator as background_estimator
-import array_analyzer.utils.io_utils as io_utils
-from array_analyzer.extract.metadata import MetaData
+import multiSero.array_analyzer.extract.image_parser as image_parser
+import multiSero.array_analyzer.extract.txt_parser as txt_parser
+import multiSero.array_analyzer.extract.img_processing as img_processing
+import multiSero.array_analyzer.load.debug_plots as debug_plots
+import multiSero.array_analyzer.load.report as report
+import multiSero.array_analyzer.extract.constants as constants
+import multiSero.array_analyzer.transform.array_generation as array_gen
+import multiSero.array_analyzer.extract.background_estimator as background_estimator
+import multiSero.array_analyzer.utils.io_utils as io_utils
+from multiSero.array_analyzer.extract.metadata import MetaData
 
 
 def interp(input_dir, output_dir):
