+++ multiSero/interpretation/od_analyzer.py
@@ -1,12 +1,11 @@
 import pandas as pd
-import numpy as np
 import os
 import re
 from matplotlib import pyplot as plt
 import seaborn as sns
-from interpretation.plotting import roc_plot_grid, total_plots
-from interpretation.report_reader import slice_df, normalize_od, read_output_batch
-import array_analyzer.extract.constants as constants
+from multiSero.interpretation.plotting import roc_plot_grid, total_plots
+from multiSero.interpretation.report_reader import slice_df, normalize_od, read_output_batch
+import multiSero.array_analyzer.extract.constants as constants
 
 
 def read_config(input_dir):
@@ -143,11 +142,11 @@ def analyze_od(input_dir, output_dir, load_report):
 
 
             plt.savefig(os.path.join(constants.RUN_PATH, 'catplot_{}.png'.format(split_suffix)),
-                                          dpi=300, bbox_inches='tight')
+                        dpi=300, bbox_inches='tight')
             if cat_param_df['zoom']:
                 g.set(ylim=(-0.05, 0.4))
                 plt.savefig(os.path.join(constants.RUN_PATH, 'catplot_zoom_{}.png'.format(split_suffix)),
-                                              dpi=300, bbox_inches='tight')
+                            dpi=300, bbox_inches='tight')
         #%% 4PL fit
         if not fit_param_df.empty:
             slice_action = fit_param_df['serum ID action']
@@ -155,6 +154,6 @@ def analyze_od(input_dir, output_dir, load_report):
             dilution_df = slice_df(df_norm_sub, slice_action, 'serum ID', fit_param_df['serum ID'])
             split_subplots_by = fit_param_df['split subplots by']
             total_plots(dilution_df, constants.RUN_PATH, 'fit_{}'.format(split_suffix), 'png', hue=hue,
-                                zoom=fit_param_df['zoom'], split_subplots_by=split_subplots_by, col_wrap=2)
+                        zoom=fit_param_df['zoom'], split_subplots_by=split_subplots_by, col_wrap=2)
         plt.close('all')
