+++ multiSero/interpretation/report_reader.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 import pandas as pd
 import logging
-import array_analyzer.extract.constants as constants
+import multiSero.array_analyzer.extract.constants as constants
 
 def antigen2D_to_df1D(xlsx_path, sheet, data_col):
     """
