+++ tests/test_debug_plots.py
@@ -1,9 +1,3 @@
-import pytest
-
-import array_analyzer.load as load
-import os
-
-
 def test_nonetype_integration():
     pass
